move SingleSelectionAttribute to manager
